requirements.txt
fastapi
uvicorn[standard]
ultralytics
python-multipart
Pillow